+++ src/main/java/myjournal/EmptyDescriptionException.java
@@ -1,16 +1,18 @@
+package myjournal;
+
 /**
- * A class to create invalid task number exceptions.
+ * A class to create empty description exceptions.
  *
  * @author Felissa Faustine
  */
-public class InvalidTaskNumberException extends RuntimeException {
+public class EmptyDescriptionException extends MyJournalException {
 
     /**
-     * The constructor for the InvalidTaskNumberExceptions.
+     * The constructor for the EmptyDescriptionExceptions.
      *
      * @param message The error message.
      */
-    public InvalidTaskNumberException(String message) {
+    public EmptyDescriptionException(String message) {
         super(message);
     }
 
@@ -21,6 +23,6 @@ public class InvalidTaskNumberException extends RuntimeException {
      */
     @Override
     public String toString() {
-        return "InvalidTaskNumberException: " + getMessage();
+        return "EmptyDescriptionException: " + getMessage();
     }
 }
