+++ src/main/java/myjournal/MyJournalException.java
@@ -1,0 +1,12 @@
+package myjournal;
+
+public class MyJournalException extends RuntimeException {
+    public MyJournalException(String message) {
+        super(message);
+    }
+
+    @Override
+    public String toString() {
+        return "MyJournalException: " + getMessage();
+    }
+}
