+++ src/main/java/myjournal/Storage.java
@@ -1,9 +1,15 @@
+package myjournal;
+
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.FileWriter;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Scanner;
+import myjournal.Task;
+import myjournal.Todo;
+import myjournal.Event;
+import myjournal.Deadline;
 
 public class Storage {
     private File file;
