+++ src/main/java/myjournal/Task.java
@@ -1,3 +1,5 @@
+package myjournal;
+
 /**
  * An abstract class for tasks.
  *
@@ -56,8 +58,9 @@ public abstract class Task {
     }
 
     /**
-     * Returns the string representation of the task;
-     * @return
+     * Returns the string representation of the task.
+     *
+     * @return The string representation of the task.
      */
     @Override
     public String toString() {
